requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
openai==1.51.0
pydantic==2.9.2
aiofiles==24.1.0
